+++ wfsource/source/baseobject/msgport.cc
@@ -62,7 +62,7 @@ MsgPort::~MsgPort()
 //============================================================================
 // Returns true if successful, false if error.
 
-bool MsgPort::PutMsg(int16 msgType, int32 msgData)
+bool MsgPort::PutMsg(int16 msgType, uintptr_t msgData)
 {
 
 #pragma message ("KTS " __FILE__ ": this needs to occur somewhere which knows about actors")
